[Files]: Metadata cleanup.
There is a large hierarchy of classes in
ui/file_manager/file_manager/foreground/js/metadata/ which seems to be
directly replaceable with standard JavaScript object. Here we are
eliminating two classes that define essentially an interface of a Map
object and a Object based implementation of it.

Bug: b/199451464
Change-Id: I2617f74a0778f4c475955f18c0f69a8cb2f1e50c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3859045
Commit-Queue: Bo Majewski <majewski@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1040226}