Refactor URLLoader creation for tests
Reduce the number of calls to the URLLoader constructor in the URLLoader
unit tests to make refactoring easier. Create a test-only
URLLoaderOptions struct which has common defaults for each option and
then override only the options which are relevant to each test.

URLLoaderOptions has a MakeURLLoader method to create a URLLoader with
the specified options. It can only be called once per URLLoaderOptions
object.

BUG=1336446

Change-Id: I4796dc2c5da83767eb0f09a691a19af3b7707639
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3855811
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1040281}