cros_healthd: service manager migration
Add a switch to migrate to chromeos mojo service manager. This flag will
make ash-chrome use service manager instead of dbus to bootstrap mojo
connection to healthd.

The reason to not just query whether the healthd exist in the service
manager is, the original bootstrap flow of dbus is sync, and changing
that to async could introduce additional issues. The flag will be
removed after healthd is fully migrated to service manager.

Bug: b:230064284, b:236953728
Change-Id: I4fec2785ee074c483d8e9ff2b53711188acafb52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3758177
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Chung-Sheng Wu <chungsheng@google.com>
Cr-Commit-Position: refs/heads/main@{#1040293}