Adds check to ensure Media Route enabled.
In response to user reported crashes in ChromeOs, the
AccessCodeCastSinkService was being created when the media router was
not enabled for that user.

This change adds an additional logic check to ensure that no ACC
elements are created if the media router is not enabled. The key change here is in access_code_cast_feature.cc where we add an additional check for MediaRouterEnabled(profile).

The rest of the changes in the code reflect the updated checking mechanism that requires a Profile and not PrefService + any changes to unittests.

Bug: b/243848498, 1356753
Change-Id: I2b8b778d71c71ad12f18d6671a442acd7f168347
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3857113
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Takumi Fujimoto <takumif@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Brian Malcolm <bmalcolm@chromium.org>
Commit-Queue: george benz <gbj@google.com>
Cr-Commit-Position: refs/heads/main@{#1040300}