Update pending_beacon's browsertest so that it's not crashed due to low memory
https://crbug.com/1356665 happens on low memory mobile device: when
navigating away to 2nd page, the 1st page is not necessary put into bfcache even if bfcache TTL is long. If the 1st page is destroyed but the test
expects it hidden, the test crashes.

This CL force BFCache to work by not setting its memory thresholds.

Bug: 1356665,1293679
Change-Id: I589c2be6f981b6b79e33222964e853f1eab128e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3858830
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Commit-Queue: Ming-Ying Chung <mych@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1040361}