[ios] Fix issue with Share Chrome on iPad
Previously the BrowserViewController's view  was being given as the
`originView` for the "share sheet". On iPad, the "share sheet" is
presented as a popup with an arrow pointing to the view that originated
it. Apparently when you use a view that is the whole screen, it just
doesn't work, probably because it can't present a popup that doesn't
overlap the origin view.

The fix was to use the tools menu button as the origin view, which
results in the popup pointing to that as the origin. Screenshot:
https://screenshot.googleplex.com/55nRMUcE9cFYosV.png

This change also adds an EarlGray test to ensure that the "share sheet"
is displayed as expected.

Bug: 1355813
Change-Id: I3f33361b6182601db96785c3cece995da42bf789
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3854651
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1040382}