[Graphics] Mark StreamTextureSII's GLImage as bound when binding
The goal of this CL is to eliminate the GL command decoders' thinking
that StreamTextureSharedImageInterface instances need lazy binding when
clients have correctly called BeginSharedImageAccessDirectCHROMIUM()
before invoking drawing operations. Currently, the SharedImageBacking
subclass that sets the StreamTextureSII instance in the texture [1] does
not mark it as bound in BeginAccess(), so the decoders are not aware
that the images have been bound.

This CL adds such marking. However, it turns out that that is not
enough: the two subclasses of StreamTextureSII currently each override
HasMutableState() to return false, which means that TextureManager
short-circuits out of actually updating its binding state for the
image. This was implemented to ensure that copies of the image are
not mistakenly elided when one of these subclasses is reused with a
different buffer. In StreamTexture, both CopyTexImage() and
BindTexImage() are effectively no-ops, meaning that the semantic change
doesn't actually matter. In CodecImage this is not the case. To ensure
that the necessary semantics here are preserved, this CL also sets
the image state to be unbound in the SharedImageBacking sublasses'
implementations of EndAccess(). We do not believe that CodecImage
will have its associated buffer changed while a client is accessing
it (i.e., in between a call to BeginAccess() and a call to EndAccess()).

Note: It is possible to see the impact of this change on lazy binding
in tests by viewing the trybot results on PS3 and PS8:

- PS3 has CHECKs inserted where the decoders do lazy binding. Android
  test failures are due to this issue.
- PS8 keeps the CHECKs and adds the production fix in this CL. All of
  the Android test failures disappear.

[1] https://source.chromium.org/chromium/chromium/src/+/main:gpu/command_buffer/service/shared_image/video_surface_texture_image_backing.cc;drc=db0d32ea0c658e4744009d0935d3da0d8b5e7b0a;l=191

Bug: 1345585
Change-Id: I3434bd82dc7b76d01872414aa7232dd870d1cd8c
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3790843
Reviewed-by: vikas soni <vikassoni@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1040435}