[UIA] Fix generated newline readonly attribute for sibling inputs
In CL:3860193, we modified the value returned for the readonly attribute
on text ranges that only contain a generated newline. I found a corner
case for which this fix didn't work:

data:text/html,<input style="display: block" type=text><input type=text>

This lays out two input nodes with a paragraph boundary in between. When
the text range starts at the end of the first input node and ends at the
beginning of the last input node, NormalizeTextRange() doesn't modify
the range at all (because HasTextRangeOrSelectionInAtomicTextField()
returns true), thus the condition where we compare the two normalized
endpoints doesn't work as expected.

To fix it, I slightly refactored my fix so that it doesn't rely on the
full normalizing function but rather only the subet of it that we need
for this case.

Change-Id: Ibfaef13af3236be13d6c4de5664dfc6ee0434371
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3861913
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1040473}