glanceables: Initial skeleton for connection with full restore
For this initial skeleton, suppress the full restore notification for
the primary user (where we show glanceables instead). This will let
PM/UX try the behavior and give us feedback on how they want it to
behave in the final version.

When the user clicks on the "restore session" button in glanceables,
trigger the FullRestoreService's session restore logic, as if the
user had clicked the "Restore" button in the notification.

Bug: 1355521
Test: added to ash_unittests, unit_tests
Change-Id: I22cb212e9fc92d9f5d01024a814845a773aff889
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3857187
Reviewed-by: Matthew Mourgos <mmourgos@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1040481}