[loader] Send cached metadata as part of OnReceiveResponse
CachedMetadata is an opaque data blob optionally stored alongside
responses in the response cache. It stores data specific to the
response -- in particular, for JavaScript files it will store the V8
code cache. Before this patch, CachedMetadata was sent to the renderer
using a separate OnReceiveCachedMetadata message, after
OnReceiveResponse.

This mechanism is only used by the Service Worker installed resources; in the HTTP cache case, there is a separate "isolated code cache"
which is queried at the same time as the HTTP cache and the two
requests are synchronised (cf. https://crbug.com/812168). This means
that, before this patch, HTTP code caches were received at the same
time as the body (thanks to the synchronisation), but Service Worker code caches were received _after_ the body (which is sent as part of
OnReceiveResponse). The consequence was that Service Worker code
caches were dropped in favour of (re)parsing the script response body,
as that parse was started before the CachedMetadata was received.

Now, rather than a separate OnReceiveCachedMetadata, send an optional
CachedMetadata alongside the headers and body in OnReceiveResponse.
This allows the receiver of the response to more carefully interleave
dealing with the metadata and body.

Review note: Since the Service Worker is the only user of this API, the
interesting parts of this patch are the handling of service worker
and the changes in the Blink UrlLoaderClient. Unfortunately, there are
other implementers of the UrlLoaderClient interface which either
ignore or trivially pass through the cached metadata, and the
signature change of the OnReceiveResponse method makes this CL
impossible to split by subdirectory for these trivial changes.

Bug: 1350077
Change-Id: I43cce62bf573691247cb8ccc362a4b73541185a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3811219
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Josh Simmons <simmonsjosh@google.com>
Cr-Commit-Position: refs/heads/main@{#1040492}