[ios] Allow Modules to be more flexible on NTP initial load
Before the Feed's CollectionView is ready and added as the parent
view, a native UICollectionView is used and can impose a height
on the ContentSuggestionsViewController's view the size of the NTP.
Giving the module container height flexibility is the simplest way
to accommodate this.

The URLDragDropHandler for ContentSuggestions is also moved to the
viewcontroller since there accessing the view before it is added
to the NTP will prematurely trigger a viewDidLoad.

Bug: 1356093
Change-Id: I559978004262c7e79739b8ed8f398b4958d2ee28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3856736
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1040520}