Shimless: fix transitions between network page and select components
The welcome page sets up the welcome state choice, which is
necessary to send the rmad request to start repair. That request
will be sent by the network page (if it's shown), because the network
page has the same rmad state as the welcome page. If the network page
is visited by going back from the next page, the welcome state choice
won't be set because the network page itself doesn't set it (only the
welcome page does). That means the next time you go forward from the
network page, the request to start repair will be sent without the
welcome state choice, which will cause an rmad error (missing arguments)
and will result in skipping back to the welcome state.

This CL fixes the bug.

Tested manually, because the bug involves a response from the platform
side.

Bug: 1198187, b/243327545
Test: local
Change-Id: If6e90d3fd1b528d11fc26daa46b864c1f677007b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3857729
Auto-Submit: Anton Swifton <swifton@google.com>
Reviewed-by: Gavin Williams <gavinwill@chromium.org>
Commit-Queue: Gavin Williams <gavinwill@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1040522}