Record LCP value at the first OnHidden in UKM and UMA
Right now, we report the LCP UKM and UMA metrics at OnComplete, which
may not be reported when the tabs are never closed or the browser
crashes. Thus, we get around 10% less of LCP reported than FCP. Since we
cannot update metrics in UKM, we need to create new LCP metric which can
fill in for the largest contentful paint values that are not able to be
recorded at tab close, when the current LCP reporting happens.

The privacy review of adding the additional LCP metric:
https://docs.google.com/document/d/1uQJlgUSeFqmG5tqlJJhmutsLHFoFeTASCiPyw9mafe0

Bug: 1210681
Change-Id: Ic3168b894c8fbabf00c71c57473b4cb4251d9ff4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3824683
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Reviewed-by: Nicolás Peña <npm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1040549}