[mparch] Fix some incorrect usages of RFHI::is_main_frame
With MPArch, we can have nested frame trees, so RFHI::is_main_frame
cannot be used for "outermost" semantics. This CL changes a number of
simpler cases where semantics match outermost. This should have no
effect outside of the use of MPArch-based nested frame trees (only
fenced frames at the moment).

Usually the replacement for RFHI::is_main_frame in these cases is
RFHI::IsOutermostMainFrame, but for a11y-related functions, the correct
function is RFHI::AccessibilityIsMainFrame.

A number of usages do not have tests included in this CL, but I have
added TODOs for cases where it seems they'd be useful (similar to the
approach taken in crrev.com/c/2988476). My heuristic for adding these
TODOs was to add them for functional changes, but to skip for tests,
logging/metrics, and DCHECKs.

shivanisha@ also pointed out that for the test RFH, we should be using
the navigation controller's ShouldMaintainTrivialSessionHistory
method, so I've also included that in this fix (also required changing
visibility of that method on NavigationControllerImpl).

This CL is based on the auditing done by mcnee@, liviutinta@,
adithyas@, vollick@ (and thanks to all who helped answer questions to
inform that work).

Bug: 1316388
Change-Id: I083a6d9fbc55cad4dc2763a8a61ca8482cae9756
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3846121
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Ian Vollick <vollick@chromium.org>
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Reviewed-by: Adithya Srinivasan <adithyas@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1040554}