Remove the use of EXPECT_DCHECK_DEATH in the browser test.
This unblocks change https://crrev.com/c/3862239.
Chatted offline with the test owner ahemery@. The reason is that
EXPECT_DCHECK_DEATH duplicates the entire process and verifies that
the process dies, which result in the DB (added in the change) is
accessed in both process and gets locked.

Bug: 1264104
Change-Id: Ib604acf25724766958cb84a199ab92787e653f07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3860314
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Shuran Huang <shuuran@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1040595}