webui: Add tests for chrome-untrusted:// service worker isolation
Adds the following tests:

1. Test that service workers and navigations to different WebUIs don't
re-use an existing WebUI service worker process.

2. Test that service workers and navigations to Web Urls don't re-use
an existing WebUI service worker process.

3. Test that new service workers and navigations to the same WebUI
don't re-use an existing WebUI service worker process. Ideally, we
would re-use a process if it was for the same WebUI but we currently
don't because of crbug.com/1158277

Also fixes an issue with ScopedWebUIConfigRegistration on Windows.
A unique_ptr was being referenced after it was moved because the
evaluation order is different on Windows.

Bug: b:242269992
Change-Id: I08b0adf1c44a1a815a35fc4a37e3ee92e853dd99
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3841531
Commit-Queue: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1040676}