wayland-protocols: Add "content-type" protocol
Source:
https://gitlab.freedesktop.org/wayland/wayland-protocols/-/merge_requests/150/diffs

This protocol is still in MR but is close to being added. Bringing it
in because it's the perfect solution for a bug we're dealing with, and
I'll update it once the MR is fully in. There are just some comment
nits atm.

BUG=b:235150518

Change-Id: Ifaef12032703c953d5ac0d815c83745e293173a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3851045
Commit-Queue: Jim Shargo <jshargo@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1040692}