Announce label as hint when the node has label for click
Android announces labels in hints to let users know what will happen if
users perform the action announced in the hint.
After this change, ChromeVox announces the label when the focused node
is clickable and has a label for click.
This change also adds a new API to get the label for long click to
prepare to announce hints for long click, too.

This design is discussed in go/label-arc

Test: AccessibilityNodeInforDataWrapperTest
Bug: b:183573519
Change-Id: I59a6b2063b2196e45ce32345c224b70b7e69891a
AX-Relnotes: hints give more information in some ARC nodes.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3692647
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Saho Kobayashi <sahok@chromium.org>
Reviewed-by: Hiroki Sato <hirokisato@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1040726}