[FilesTrash] Update TrashBanner to new icon and button
The UI spec requires the Trash banner to have the trash can icon and the
trash banner to include a button to "Empty trash now" [1]. Further the
strings updated to align with the strings at go/cros-trash-strings.

Some of the tests relied on the old "Empty Trash" button in the
toolbar so this updates those tests as well.

[1] https://docs.google.com/presentation/d/1Ipqfd5tnVXKyK-Xdf7BJzsA42VAjQEZU_-8QSqg1Yfo/edit#slide=id.g13c9d0cebe0_0_5

Bug: b:243879041
Test: browser_tests --gtest_filter=*trash*:*tabindex*:*dirContext*
Change-Id: I5ca55c36c6f6b8eca081519abfcf4ff24ff34a1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3864403
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1040823}