gin: Remove broken DCHECKs in DumpHeapStatistics
DumpHeapStatistics invokes:
  1) V8's GetHeapStatistics() and then
  2) GetHeapSpaceStatistics() for each individual space.

The DCHECKs ensures that the total sizes reported for
GetHeapStatistics() matches the sum of the individual spaces. Due
to concurrent allocation spaces can grow in-between operation 1)
and 2), so the DCHECKs allow for that as well. However some metrics
might even decrease between 1) and 2).

total_physical_size() might decrease due to sweeping, while
used_heap_size() might decrease in the concurrent allocation path. So
both values can't be used for these DCHECKs.

However, we should be able to keep the DCHECK for total_heap_size().
This value is only decreased during a GC when we release pages, which
can't happen in-between 1) and 2) here.

While it's unfortunate that we have to drop those two DCHECKs, the
remaining DCHECK will still ensure that all spaces are properly
accounted.

Bug: chromium:1357181
Change-Id: I4c7a74c4bac8fcd517858944509d8589ed7ce9e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3864664
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1040841}