isolated-apps: Add reader for Isolated Web Apps
This CL introduces a new `SignedWebBundleReader` class, which is similar
to `content::WebBundleReader`, but extended to also support reading the
integrity block and (as part of an upcoming CL) verify the integrity.

The class currently takes a `base::FilePath` to designate the location
of the web bundle. In the future, we might need a different way to
designate the location when we support reading from APKs or other
container formats.

Public Explainer: https://github.com/reillyeon/isolated-web-apps
Design Docs: go/isolated-apps-serving, go/isolated-app-scheme

Bug: 1334594
Change-Id: Idfe37fea27bd1fd275953ad03a02b2672ba55ab7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3738600
Reviewed-by: Kunihiko Sakamoto <ksakamoto@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Oleksandr Peletskyi <peletskyi@chromium.org>
Reviewed-by: Phillis Tang <phillis@chromium.org>
Commit-Queue: Christian Flach <cmfcmf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1040859}